Moved mysql.py from util to db
