+++ app/api/v1/endpoints/auth.py
@@ -8,7 +8,7 @@ from app.schemas.auth import TokenRequest, Token
 from app.schemas.user import UserOut, UserCreate
 from app.utils.audit.actions import log_action
 from app.utils.auth.actions import perform_action_auth
-from app.utils.db.mysql import get_db
+from app.db.mysql import get_db
 
 router = APIRouter()
